requirements.txt

openai==0.28.0
praw==7.8.1
python-dotenv==1.1.1
